refactor: Apply Type-Driven Design to HeroAttributes and GameState
  - Replace Int16 with Attribute wrapper type in HeroAttributes
  - Replace raw Int action points with ActionPoints type in GameState
  - Add type-safe value types: Attribute, ActionPoints, HitPoints, etc.
  - Update all dependent services, views and tests

+++ Packages/elf_Kit/Tests/elf_KitTests/DataLayer/Services/Armor/ElfArmorServiceTests.swift
@@ -175,7 +175,7 @@ final class ElfArmorServiceTests: XCTestCase {
         let service = ElfArmorService(itemsRepository: repository)
         let result = await service.getAllItemsArmor(for: [id])
 
-        XCTAssertEqual(result.values.reduce(0, +), 0)
+        XCTAssertEqual(result.values.reduce(Int16(0), +), 0)
     }
 
     func testMissingItemIsIgnored() async throws {
@@ -185,6 +185,6 @@ final class ElfArmorServiceTests: XCTestCase {
         let service = ElfArmorService(itemsRepository: repository)
         let result = await service.getAllItemsArmor(for: [id])
 
-        XCTAssertEqual(result.values.reduce(0, +), 0)
+        XCTAssertEqual(result.values.reduce(Int16(0), +), 0)
     }
 }
